+++ core/src/main/kotlin/com/tinder/gitquery/core/config/GitQueryConfigSchema.kt
@@ -1,0 +1,12 @@
+/*
+ * © 2019 Match Group, LLC.
+ */
+
+package com.tinder.gitquery.core.config
+
+/**
+ * A model fof the schema attribute of the config file.
+ */
+data class GitQueryConfigSchema(
+    var version: String = "1"
+)