+++ gradle-plugin/src/main/kotlin/com/tinder/gitquery/GitQueryPlugin.kt
@@ -12,8 +12,16 @@ import org.gradle.api.Project
  */
 class GitQueryPlugin : Plugin<Project> {
     override fun apply(project: Project) {
-        val extension =
-            project.extensions.create("gitQuery", GitQueryExtension::class.java, project)
-        project.tasks.register("gitQuery", GitQueryTask::class.java, extension)
+        val syncExtension = project.extensions.create("gitQuery", GitQuerySyncExtension::class.java, project)
+        val initExtension = project.extensions.create("gitQueryInit", GitQueryInitExtension::class.java, project)
+
+        project.tasks.register("gitQuery", GitQuerySyncTask::class.java, syncExtension)
+        project.tasks.register("gitQueryInit", GitQueryInitTask::class.java, syncExtension, initExtension)
+
+        project.afterEvaluate {
+            if (syncExtension.autoSync) {
+                project.tasks.getByName("assemble").dependsOn("gitQuery")
+            }
+        }
     }
 }